refactor: Move the classes to more fitting packages
