+++ ors-benchmark/src/test/java/org/heigit/ors/util/ProgressBarLoggerTest.java
@@ -3,7 +3,6 @@ package org.heigit.ors.util;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
-import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.ConsoleAppender;
